BUGFIX: Feeds were not being properly fetched/overwritten. (#3)
In times of high stress, multiple pushes get made at the same time to the DB. To combat this, the Feeds DB has a complex unique index on it. (Which still needs to be added to the appropriate Feeds schema, like so: Automattic/mongoose#3393)

This patch resolves part of the issues currently being experienced on the server.